refactor(cli): split verify exec helpers
Move verify build and subprocess helpers into runtime_exec.go so runtime.go stays focused on verification orchestration. Collapse runCLI onto runCLIWithOutput to keep command execution semantics in one place while preserving caller behavior.

Verification:
- go test ./internal/pipeline
- go test ./...
- go vet ./...
- golangci-lint run ./...
- scripts/golden.sh verify
- go build -o ./printing-press ./cmd/printing-press
